+++ CHANGELOG.md
@@ -61,7 +61,7 @@
 - Remove dead and unreachable compiler error and warning variants; add fixtures for the ones found to be reachable. https://github.com/rescript-lang/rescript/pull/8459
 - Convert OCaml codebase to snake case format. https://github.com/rescript-lang/rescript/pull/8456
 - Analysis refactor: remove global state `Shared_types.state`. https://github.com/rescript-lang/rescript/pull/8465
-
+- Refactor analysis CLI helpers to use source input. https://github.com/rescript-lang/rescript/pull/8466
 
 # 13.0.0-alpha.4